+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapServer.lua
@@ -221,7 +221,7 @@ local function S7_Config_ModMenuRelay(Signal) --  Signal recieved from Osiris.
         end
     end
 
-    S7_SetDialogVars() --  Request dialogVar update everytime ModMenu relays a signal.
+    SetDialogVars() --  Request dialogVar update everytime ModMenu relays a signal.
 end
 
 --  ============================================================================